requirements.txt
pytest
pytest-xdist
allure-pytest
playwright
